refactor(extension): deduplicate wallet activities fetching into single hook
